refactor(recipes): refactor recipe controller, repository and routes

+++ src/app/repositories/RecipesRepository.ts
@@ -2,13 +2,18 @@ import {Recipes} from '@prisma/client'
 import {prisma} from '../database'
 
 class RecipesRepository {
-  async findAll(): Promise<Recipes[]> {
-    return prisma.recipes.findMany()
+  async findAll(userId: string): Promise<Recipes[]> {
+    return prisma.recipes.findMany({
+      where: {userId},
+    })
   }
 
-  async findById(id: string): Promise<Recipes | null> {
-    return prisma.recipes.findUnique({
-      where: {id},
+  async findById(id: string, userId: string): Promise<Recipes | null> {
+    return prisma.recipes.findFirst({
+      where: {
+        id: id,
+        userId: userId,
+      },
     })
   }
 
@@ -16,29 +21,27 @@ class RecipesRepository {
     return prisma.recipes.create({
       data: {
         ...recipe,
-        user: {
-          connect: {id: userId},
-        },
+        userId,
       },
     })
   }
 
-  async update(id: string, recipe: Partial<Recipes>): Promise<Recipes | null> {
+  async update(id: string, userId: string, recipe: Partial<Recipes>): Promise<Recipes | null> {
+    const recipeToUpdate = await this.findById(id, userId)
+    if (!recipeToUpdate) return null
+
     return prisma.recipes.update({
-      where: {id},
+      where: {id: recipeToUpdate?.id},
       data: recipe,
     })
   }
 
-  async delete(id: string): Promise<Recipes | null> {
-    return prisma.recipes.delete({
-      where: {id},
-    })
-  }
+  async delete(id: string, userId: string): Promise<Recipes | null> {
+    const recipeToDelete = await this.findById(id, userId)
+    if (!recipeToDelete) return null
 
-  async findByUserId(userId: string): Promise<Recipes[]> {
-    return prisma.recipes.findMany({
-      where: {userId},
+    return prisma.recipes.delete({
+      where: {id: recipeToDelete?.id},
     })
   }
 }
